Add option to save synchronize log to logfile
Introduced a new state variable `saveactualsynclogdata` to control saving the output of real synchronize tasks to a logfile. Updated SidebarMainView, SidebarTasksView, and TasksView to propagate this state, and added a toolbar button in TasksView to toggle this option. Also added a message view to indicate when log saving is active.